Move hymod_fluxes struct to dedicated file.
