+++ models/hymod/include/Hymod.h
@@ -9,35 +9,12 @@
 #include "Pdm03.h"
 #include "hymod_params.h"
 #include "hymod_state.h"
-
-//! Hymod flux structure
-/*!
-    This structure provides storage for the fluxes generated by hymod at any time step
-*/
-
-struct hymod_fluxes
-{
-    double slow_flow_meters_per_second;  //!< The flow exiting slow flow at this time step
-    double runoff_meters_per_second;     //!< The caluclated runoff amount for this time step
-    double et_loss_meters;    //!< The amount of water lost to evapotranspiration
-
-    //! Constructor for hymod fluxes
-    /*!
-        Default constructor for hymod fluxes this is needed for the structs to be storeable in C++ containers.
-    */
-
-    hymod_fluxes(double sf = 0.0, double r = 0.0, double et = 0.0) :
-        slow_flow_meters_per_second(sf), runoff_meters_per_second(r), et_loss_meters(et)
-    {}
-};
+#include "hymod_fluxes.h"
 
 //! Hymod kernel class
 /*!
     This class implements the hymod hydrological model
-
-
 */
-
 class hymod_kernel
 {
     public:
